+++ src/main/java/be/podor/security/WebSecurityConfig.java
@@ -5,12 +5,10 @@ import be.podor.security.jwt.JwtConfiguration;
 import be.podor.security.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.security.SecurityProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
-import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
@@ -31,13 +29,11 @@ public class WebSecurityConfig {
     @Autowired
     private final JwtTokenProvider jwtTokenProvider;
 
-//    private final AccessDeniedHandlerException accessDeniedHandlerException;
-//
-//    private final AuthenticationEntryPointException authenticationEntryPointException;
     @Bean
     public PasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
     }
+
     @Bean
     @Order(SecurityProperties.BASIC_AUTH_ORDER)
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
@@ -47,30 +43,19 @@ public class WebSecurityConfig {
 //        csrf 비활성화
         http.csrf().disable()
 
-//               예외 처리
-                .exceptionHandling()
-//                .authenticationEntryPoint(authenticationEntryPointException)
-//                .accessDeniedHandler(accessDeniedHandlerException)
 
 //                세션 미사용 설정
-                .and()
                 .sessionManagement()
                 .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
 
 //                api 허용 목록!
                 .and()
                 .authorizeRequests()
-                .antMatchers(HttpMethod.GET, "/api/**").permitAll()
-                .antMatchers("/member/login").permitAll()
-                .antMatchers("/api/register").permitAll()
-                .antMatchers("/swagger-ui.html/**").permitAll()
                 .anyRequest().permitAll()
                 .and()
-                .oauth2Login()
 
 
 //                필터 적용
-                .and()
                 .apply(new JwtConfiguration(jwtTokenProvider));
 
         return http.build();
